+++ CHANGELOG.md
@@ -8,9 +8,13 @@
 
 - Add optional seconds parameter to the `clear` method of the Cache system.
 
+- Add support for dynamic redirections to the Route system.
+
 - Add better overall performance.
 
 ### Changed
+
+- Change the function parameter of the `Wolff\Core\Route` from an optional string to an optional array of strings containing the controller and method to call (`'home@index'` now is `[ Controller\Home::class, 'index' ]`).
 
 - Make the Logging system non-static.